requirements.txt
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
archspec @ file:///croot/archspec_1709217642129/work
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1720453481653/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
conda @ file:///croot/conda_1722004606466/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1721662679737/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
coverage==7.6.3
cryptography @ file:///croot/cryptography_1714660666131/work
distro @ file:///croot/distro_1714488253808/work
flake8==7.1.1
Flask==2.0.3
frozendict @ file:///croot/frozendict_1713194832637/work
greenlet==3.1.1
idna @ file:///croot/idna_1714398848350/work
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
MarkupSafe==3.0.1
mccabe==0.7.0
menuinst @ file:///croot/menuinst_1723567589013/work
packaging @ file:///croot/packaging_1720101850331/work
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
pycodestyle==2.12.1
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes==3.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dotenv==1.0.1
requests @ file:///croot/requests_1721410876868/work
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
setuptools==72.1.0
SQLAlchemy==1.4.22
tqdm @ file:///croot/tqdm_1716395931952/work
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
urllib3 @ file:///croot/urllib3_1718912636303/work
Werkzeug==2.0.3
wheel==0.43.0
zstandard @ file:///croot/zstandard_1714677652653/work
